requirements.txt
google-generativeai==0.5.4

moviepy
ffmpeg-python
pydub

opencv-python
Pillow

sentence-transformers==2.2.2
transformers==4.33.2
huggingface-hub==0.20.3
sentencepiece

TTS==0.21.3

audioread
soundfile
soxr

numpy
scipy
pandas
scikit-learn

PyYAML
requests
tqdm
protobuf
